+++ container/raw/factory.go
@@ -15,6 +15,7 @@
 package raw
 
 import (
+	"flag"
 	"fmt"
 
 	"github.com/golang/glog"
@@ -23,6 +24,8 @@ import (
 	"github.com/google/cadvisor/fs"
 	info "github.com/google/cadvisor/info/v1"
 )
+
+var dockerOnly = flag.Bool("docker_only", false, "Only report docker containers in addition to root stats")
 
 type rawFactory struct {
 	// Factory for machine information.
@@ -33,6 +36,9 @@ type rawFactory struct {
 
 	// Information about mounted filesystems.
 	fsInfo fs.FsInfo
+
+	// Watcher for inotify events.
+	watcher *InotifyWatcher
 }
 
 func (self *rawFactory) String() string {
@@ -40,12 +46,30 @@ func (self *rawFactory) String() string {
 }
 
 func (self *rawFactory) NewContainerHandler(name string) (container.ContainerHandler, error) {
-	return newRawContainerHandler(name, self.cgroupSubsystems, self.machineInfoFactory, self.fsInfo)
+	return newRawContainerHandler(name, self.cgroupSubsystems, self.machineInfoFactory, self.fsInfo, self.watcher)
 }
 
-// The raw factory can handle any container.
-func (self *rawFactory) CanHandle(name string) (bool, error) {
-	return true, nil
+// The raw factory can handle any container. If --docker_only is set to false, non-docker containers are ignored.
+func (self *rawFactory) CanHandleAndAccept(name string) (bool, bool, error) {
+	accept := name == "/" || !*dockerOnly
+	return true, accept, nil
+}
+
+func (self *rawFactory) DebugInfo() map[string][]string {
+	out := make(map[string][]string)
+
+	// Get information about inotify watches.
+	watches := self.watcher.GetWatches()
+	lines := make([]string, 0, len(watches))
+	for containerName, cgroupWatches := range watches {
+		lines = append(lines, fmt.Sprintf("%s:", containerName))
+		for _, cg := range cgroupWatches {
+			lines = append(lines, fmt.Sprintf("\t%s", cg))
+		}
+	}
+	out["Inotify watches"] = lines
+
+	return out
 }
 
 func Register(machineInfoFactory info.MachineInfoFactory, fsInfo fs.FsInfo) error {
@@ -57,11 +81,17 @@ func Register(machineInfoFactory info.MachineInfoFactory, fsInfo fs.FsInfo) erro
 		return fmt.Errorf("failed to find supported cgroup mounts for the raw factory")
 	}
 
+	watcher, err := NewInotifyWatcher()
+	if err != nil {
+		return err
+	}
+
 	glog.Infof("Registering Raw factory")
 	factory := &rawFactory{
 		machineInfoFactory: machineInfoFactory,
 		fsInfo:             fsInfo,
 		cgroupSubsystems:   &cgroupSubsystems,
+		watcher:            watcher,
 	}
 	container.RegisterContainerHandlerFactory(factory)
 	return nil